Move method to RoomController
